Fix two references to renamed module
